Move `starters.ts`
